refactor(output): extract user tests to sibling file
